libflash/ipmi-hiomap: Improve event handling
[ Upstream commit 34cffed2ccf3971b73b27245d2d11c135f67bcd8 ]

The host firmware side of the hiomap protocol has two input sources:

1. Requests to adjust the flash mappings from itself or the kernel
2. State change events received from the BMC

The handling of BMC state change events (2.) is asynchronous in two ways:

a. The BMC pushes the state change event to the host, which is recorded
   but not acted on
b. When handling requests to adjust the flash mapping, skiboot first
   addresses any new BMC state changes before servicing the mapping
   request

Further, the hiomap protocol sends a mix of ackable and stateful events,
where ackable events are only relevant until skiboot's hiomap event
handler (b. above) cleans them up, whereas stateful events persist until
the BMC provides a subsequent state change event.

As we handle the ackable events asynchronous to receiving notification
(b. vs a. above), OR in the received event state rather than directly
assign to ensure we don't lose events that we must not miss. As an
example, without the OR we may lose ackable events if the daemon
restarts and pushes a new state change event during initialisation,
which will necessarily bear no relation to the previous state change
event value.

Similarly, don't close active windows in a. based on the event content,
as we need the window type information to handle state restoration in b.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>