gcov: link in ctors* as newer GCC doesn't group them all
It seems that newer toolchains get us multiple ctors sections to link in
rather than just one. If we discard them (as we were doing), then we
don't have a working gcov build (and we get the "doesn't look sane"
warning on boot).

So, include ctors* and all is well.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>