hiomap: quieten warning on failing to move a window
[ Upstream commit 606a5a3d44e3271e5177bb925d043c8ce48cdaa3 ]

This isn't *necessarily* an error that we should complain loudly about.
If, for example, the BMC enforces the Read Only flag on a FFS partition,
opening a write window *should* fail, and we do indeed test this in
op-test.

Thus we deal with the error in a well known path: returning an error
code and then it's eventually a userspace problem.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>
Reviewed-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>