platform/witherspoon: Fix opencapi lane-mask used on GPU0
When an opencapi device is used via the Acorn adapter, the link used
is connected to the "middle" group of lanes of the obus. We were using
the wrong set of lanes. The link was somehow still training, likely
because the default settings at power-on were good enough, but it's
still wrong.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>