ipmi: Reduce ipmi_queue_msg_sync() polling loop time to 10ms
[ Upstream commit 28bf1faf44cd6bffcf3df0c00578b279f3faeab1 ]

On a plain boot, this reduces the time spent in OPAL by ~170ms on
p9dsu. This is due to hiomap (currently) using synchronous IPMI
messages.

It will also *significantly* reduce latency on runtime flash
operations, as we'll spend typically 10-20ms in OPAL rather than
100-200ms. It's not an ideal solution to that, but it's a quick
and obvious win for jitter.

Cc: stable
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>