hdata/test: workaround dtc bugs
[ Upstream commit b547df61bdadbacde6ff94099a2f89e4f167c9a8 ]

In dtc v1.4.5 to at least v1.4.7 there have been a few bugs introduced
that change the layout of what's produced in the dts. In order to be
immune from them, we should use the (provided) dtdiff utility, but we
also need to run the dts we're diffing against through a dtb cycle in
order to ensure we get the same format as what the hdat_to_dt to dts
conversion will.

This fixes a bunch of unit test failures on the version of dtc shipped
with recent Linux distros such as Fedora 29.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>