libpore: Sync p8 files, remove erroneous "IBM Confidential"
We also had some rogue "IBM Confidential" strings that we failed to
remove with the original change of Copyright headers for open sourcing.
Do this by synchronising with the hostboot copy of the code, which
removed the Confidential string when their copyright headers changed for
initial open sourcing of the code back in 2014. See hostboot commit
3bcf5b7982bb8a2d9227dbff7be4ff2ce5fec05c where the HWP copyright headers
were updated.

We likely missed this as we did a similar process inside the skiboot
repository, but likely only on the (C) headers themselves.

The libpore changes that we were missing *look* minor, but we need to
throw some testing at them at least, as there *are* changes that we were
missing.

We also have to make a minor modification (being sent upstream) to avoid
a compiler warning of always false comparison (<0 on unsigned int)

Reported-by: Dawn Sylvia <ddzubak@us.ibm.com>
Reviewed-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>