ffspart: Add test for eraseblock size
This test checks that the partitions are correctly laid out when the
eraseblock size is greater than the start of the first partition.
Currently ffspart fails to create a valid image in this case.

There are two tests. The second is expected to fail but it is marked as
passing for now.

This test requires pflash to work. Currently we leave that as an
exercise for the user.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>