pflash: Support encoding/decoding ECC'd partitions
With the new --ecc option, pflash can add/remove ECC when
reading/writing flash partitions protected by ECC.

This is *not* flawless with current PNORs out in the wild though, as
they do not typically fill the whole partition with valid ECC data, so
you have to know how big the valid ECC'd data is and specify the size
manually. Note that for some partitions this is pratically impossible
without knowing the details of the content of the partition.

A future patch is likely to introduce an option to "stop reading data
when ECC starts failing and assume everything is okay rather than error
out" to support reading the "valid" data from existing PNOR images.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>