ci: Bump Qemu version
This moves the qemu version to qemu-powernv-for-skiboot-7 which is based
on upstream's 3.1.0, and supports a Power9 machine.

It also includes a fix for the skiboot XSCOM errors:

 XSCOM: read error gcid=0x0 pcb_addr=0x1020013 stat=0x0

There is no modelling of the xscom behaviour but the reads/writes
now succeed which is enough for skiboot to not error out.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>