doc: Adjusting tags in release notes to eliminate global conflict for singlehtml builds
Change tags shared-slot-rn and capi2-rn tags to reduce conflict tags in
skiboot-5.7.rst

For the singlehtml builds to work correctly, the same tags cannot be used
globally in any files.  These two tags, shared-slot-rn and capi2-rn, also
appear in skiboot-5.7.rst. To eliminate, the
tags had a suffix "-rc1" for both the definition and use of each tag.

Signed-off-by: Jeff Scheel <scheel@us.ibm.com>
[stewart: use 5.7 in tag just to be complete]
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>