astbmc: Try IPMI HIOMAP for P8 (again)
[ Upstream commit c826e1ca9e5b8a21989325536f2f9eb5f2c87607 ]

The HIOMAP protocol was developed after the release of P8 in preparation
for P9. As a consequence P9 always uses it, but it has rarely been
enabled for P8. P8DTU has recently added IPMI HIOMAP support to its BMC
firmware, so enable its use in skiboot with P8 machines. Doing so
requires some rework to ensure fallback works correctly as in the past
the fallback was to mbox, which will only work for P9.

Tested on Garrison, Palmetto without HIOMAP, Palmetto with HIOMAP, and
Witherspoon.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>
Signed-off-by: Vasant Hegde <hegdevasant@linux.vnet.ibm.com>