npu2/hw-procedures: Fix parallel zcal for opencapi
For opencapi, we currently do impedance calibration when initializing
the PHY for the device, which could run in parallel if we were rich
and had multiple opencapi devices. But if 2 devices are on the same
obus, the 2 calibration sequences could overlap, which likely yields
bad results and is useless anyway since it only needs to be done once
per obus.

This patch splits the opencapi PHY reset in 2 parts:
- a 'init' part called serially at boot. That's when zcal is done. If
  we have 2 devices on the same socket, the zcal won't be redone,
  since we're called serially and we'll see it has already be done for
  the obus
- a 'reset' part called during fundamental reset as a prereq for link
  training. It does the PHY setup for a set of lanes and the dccal.

The PHY team confirmed there's no dependency between zcal and the
other reset steps and it can be moved earlier.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>
Reviewed-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>